+++ src/Phan/Analysis/PostOrderAnalysisVisitor.php
@@ -2048,7 +2048,7 @@ class PostOrderAnalysisVisitor extends AnalysisVisitor
         // Rarely, a conditional will always be true or always be false.
         if ($cond_truthiness !== null) {
             // TODO: Add no-op checks in another PR, if they don't already exist for conditional.
-            if ($cond_truthiness === true) {
+            if ($cond_truthiness) {
                 // The condition is unconditionally true
                 yield from $this->getReturnTypes($context, $true_node, $node->lineno);
                 return;