+++ tests/plugin_test/.phan/config.php
@@ -149,5 +149,6 @@ return [
         'DuplicateConstantPlugin',
         'ConstantVariablePlugin',
         '../../.phan/plugins/ShortArrayPlugin.php',
+        '../../.phan/plugins/SimplifyExpressionPlugin.php',
     ],
 ];